Move utils.js to utils/utils.js.
