+++ packages/ckeditor5-table/src/commands/mergecellscommand.js
@@ -10,7 +10,7 @@
 import Command from '@ckeditor/ckeditor5-core/src/command';
 import TableUtils from '../tableutils';
 import { findAncestor, updateNumericAttribute } from './utils';
-import { isSelectionRectangular, getSelectedTableCells } from '../utils';
+import { isSelectionRectangular, getSelectedTableCells } from '../utils/utils';
 
 /**
  * The merge cells command.
